% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{imm}
\alias{imm}
\title{Example Dataset: imm}
\format{
A data frame with species information.
}
\usage{
imm
}
\description{
A dataset containing species data for validation.
}
\keyword{datasets}
