% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{tab.34}
\alias{tab.34}
\title{Example Dataset: tab34}
\format{
A data frame with species information.
}
\usage{
tab.34
}
\description{
A dataset containing species information for Tab34.
}
\keyword{datasets}
