% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{tab.35}
\alias{tab.35}
\title{Example Dataset: tab35}
\format{
A data frame with species information.
}
\usage{
tab.35
}
\description{
A dataset containing species information for Tab35.
}
\keyword{datasets}
