% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{tab.36}
\alias{tab.36}
\title{Example Dataset: tab36}
\format{
A data frame with species information.
}
\usage{
tab.36
}
\description{
A dataset containing species information for Tab36.
}
\keyword{datasets}
