% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{validate.tab34}
\alias{validate.tab34}
\title{Validate Tab34 Conditions}
\usage{
validate.tab34(x, y, plot = "")
}
\arguments{
\item{x}{A dataset containing species data.}

\item{y}{A dataset containing species information for Tab34.}

\item{plot}{A string specifying the plot to analyze.}
}
\value{
A list containing the validation results.
}
\description{
This function checks if the species data in a plot meets the conditions specified for Tab34.
}
