% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{validate.tab36}
\alias{validate.tab36}
\title{Validate Tab36 Conditions}
\usage{
validate.tab36(x, y, plot = "", treshold = 9)
}
\arguments{
\item{x}{A dataset containing species data.}

\item{y}{A dataset containing species information for Tab34.}

\item{plot}{A string specifying the plot to analyze.}

\item{treshold}{A treshold value (9 or 12).}
}
\value{
A list containing the validation results.
}
\description{
This function checks if the species data in a plot meets the conditions specified for Tab36.
}
